move filter_map to its own module
